Implement circuit mixin
